feat: Add Elasticsearch config, demo guide, frontend, and update test configurations
Added ElasticsearchConfig for better ES integration setup.
Created comprehensive DEMO-GUIDE.md and demo-test.bat for testing.
Added frontend directory with UI components.
Updated security config, auth controller, and data initialization service.
Enhanced test coverage with updated application properties and test cases.

+++ src/test/java/com/globalsearch/performance/ConcurrentSearchLoadTest.java
@@ -482,7 +482,7 @@ public class ConcurrentSearchLoadTest {
 
         // Allow degraded performance during spike (up to 1000ms)
         assertThat(stats.getAverage()).isLessThan(1000);
-        assertThat(successCount.get()).isGreaterThan(spikeUsers * 0.95); // 95% success rate
+        assertThat(successCount.get()).isGreaterThan((int)(spikeUsers * 0.95)); // 95% success rate
 
         System.out.println("✓ Spike load test PASSED");
     }
